Francisco ronci branch

+++ TemplateTPIntegrador/Datos/AltaProducto.cs
@@ -12,10 +12,10 @@ namespace Datos
         string _idUsuario;
         string _idProveedor;
         string _nombre;
-        int _precio;
+        double _precio;
         int _stock;
 
-        public AltaProducto(int idCategoria, string idUsuario, string idProveedor, string nombre, int precio, int stock)
+        public AltaProducto(int idCategoria, string idUsuario, string idProveedor, string nombre, double precio, int stock)
         {
             _idCategoria = idCategoria;
             _idUsuario = idUsuario;
@@ -29,7 +29,7 @@ namespace Datos
         public string IdUsuario { get => _idUsuario; set => _idUsuario = value; }
         public string IdProveedor { get => _idProveedor; set => _idProveedor = value; }
         public string Nombre { get => _nombre; set => _nombre = value; }
-        public int Precio { get => _precio; set => _precio = value; }
+        public double Precio { get => _precio; set => _precio = value; }
         public int Stock { get => _stock; set => _stock = value; }
     }
 }
